iox-#1196 fix compiler warnings in hoofs tests

+++ iceoryx_hoofs/test/moduletests/test_cxx_optional.cpp
@@ -508,7 +508,7 @@ struct TestStructForInPlaceConstruction
     }
 
     TestStructForInPlaceConstruction(int8_t&& val)
-        : val(DEFAULT_MULTIPLICATOR * val)
+        : val(static_cast<int8_t>(DEFAULT_MULTIPLICATOR * val))
     {
     }
